+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # CHANGELOG
 
 ## Unreleased
+* Add aliases support.
 
 ## v0.5.0
 ### Added